+++ controllers/noteController.js
@@ -1,4 +1,5 @@
 const Note = require("../models/noteModel");
+const { UserRole } = require("../enums/UserRole");
 const mongoose = require("mongoose");
 
 const getNotes = async (req, res) => {
@@ -33,6 +34,12 @@ const getSpecificNote = async (req, res) => {
   }
 
   try {
+    const proceed = await canProceed(req.user._id, id);
+
+    if (!proceed && req.user.role !== UserRole.Admin) {
+      return res.status(401).json({ error: "This user is not eligible to this action" });
+    }
+
     const note = await Note.findById(id);
 
     if (!note) {
@@ -73,14 +80,20 @@ const getNotesByCriteria = async (req, res) => {
 };
 
 const createNote = async (req, res) => {
-  const { header, content, bookId, authorId, chapterId, orderId } = req.body;
+  const { header, content, bookId, chapterId, orderId } = req.body;
 
   try {
+    const proceed = await canProceed(req.user._id, id);
+
+    if (!proceed && req.user.role !== UserRole.Admin) {
+      return res.status(401).json({ error: "This user is not eligible to this action" });
+    }
+    
     const note = await Note.create({
       header,
       content,
       bookId,
-      authorId,
+      authorId: req.user._id,
       chapterId,
       orderId,
     });
@@ -96,6 +109,12 @@ const deleteNote = async (req, res) => {
   const { id } = req.params;
 
   try {
+    const proceed = await canProceed(req.user._id, id);
+
+    if (!proceed && req.user.role !== UserRole.Admin) {
+      return res.status(401).json({ error: "This user is not eligible to this action" });
+    }
+
     const deletedNote = await Note.findByIdAndDelete(id);
 
     return res.status(200).json(deletedNote);
@@ -109,6 +128,11 @@ const updateNote = async (req, res) => {
   const { header, content } = req.body;
 
   try {
+    const proceed = await canProceed(req.user._id, id);
+
+    if (!proceed && req.user.role !== UserRole.Admin) {
+      return res.status(401).json({ error: "This user is not eligible to this action" });
+    }
     const note = await Note.findOneAndUpdate(
       { _id: id },
       {
@@ -149,6 +173,16 @@ const getAllNoteTitles = async (req, res) => {
     return res.status(400).json({ error: error.message });
   }
 };
+
+async function canProceed(userId, noteId) {
+  const note = await Note.findById(noteId).select("authorId");
+
+  if (note.authorId.toString() !== userId.toString()) {
+    return false;
+  }
+
+  return true;
+}
 
 module.exports = {
   getNotes,
